requirements.txt
ccxt
praw
wheel
elasticsearch
nltk
textblob
kafka-python
pyspark
python-dotenv